feat: contribute page, middleware for side-pages
